# docs/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "DecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "description": "The named human who decided.",
            "minLength": 2,
            "title": "Actor",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionValue"
          },
          "reason": {
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "role": {
            "description": "Must appear in the stage authority matrix.",
            "minLength": 2,
            "title": "Role",
            "type": "string"
          },
          "signature": {
            "minLength": 8,
            "title": "Signature",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "role",
          "decision",
          "reason",
          "signature"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "DecisionResult": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "guardrail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardrail"
          },
          "principal": {
            "description": "The authenticated identity written into the ledger.",
            "title": "Principal",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "status",
          "detail",
          "principal"
        ],
        "title": "DecisionResult",
        "type": "object"
      },
      "DecisionValue": {
        "enum": [
          "approve",
          "reject"
        ],
        "title": "DecisionValue",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageOut"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          },
          "workflow_name": {
            "title": "Workflow Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "region",
          "status",
          "synthetic",
          "organization_id",
          "workflow_id",
          "created_at",
          "organization_name",
          "workflow_name",
          "stages"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "StageOut": {
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "allowed_roles": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Roles",
            "type": "array"
          },
          "decision": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "draft": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "escalation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Reason"
          },
          "evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "risk_tier": {
            "title": "Risk Tier",
            "type": "string"
          },
          "stage_key": {
            "title": "Stage Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stage_key",
          "name",
          "position",
          "action_type",
          "risk_tier",
          "status",
          "allowed_roles"
        ],
        "title": "StageOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "medo433447@gmail.com",
      "name": "Mohamed Mahmoud"
    },
    "description": "\nCall the governed workflow engine from another system.\n\n**What this API will not do.** It will not approve anything on your behalf: a MEDIUM or\nHIGH stage moves only through `POST /runs/{run_id}/stages/{stage_key}/decision`, carrying a\nsigned human decision, and the authenticated principal is written into the hash-chained\nledger beside it. It will also not let a caller choose the model adapter \u2014 that is a\ndeployment setting (invariant **G7**).\n\n**Authentication.** Send `Authorization: Bearer <credential>`, where the credential is\neither a scoped API key (`atmpl_...`, minted with `atmpl keys create`) or a short-lived\naccess token from `POST /oauth/token`. Scopes are enforced per route and listed in each\nroute's errors.\n\n**Errors.** Every failure is `{\"error\": {\"code\", \"message\", \"details\"}}`. Every response\nechoes the `X-Request-Id` you sent, or one generated for you.\n\nAll demo organizations, people, and records reachable through this API are synthetic.\n",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Enterprise AI Automation Templates API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs/{run_id}": {
      "get": {
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "description": "Run id.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "Run id.",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one run with its stages",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/runs/{run_id}/stages/{stage_key}/decision": {
      "post": {
        "description": "The only path that moves a MEDIUM or HIGH stage. The authenticated principal is written into the hash-chained ledger next to the signature, so the audit answers *who decided*, not only *what was decided*.",
        "operationId": "submit_decision_runs__run_id__stages__stage_key__decision_post",
        "parameters": [
          {
            "description": "Run id.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "Run id.",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Stage key inside that run, e.g. `terminal_decision`.",
            "in": "path",
            "name": "stage_key",
            "required": true,
            "schema": {
              "description": "Stage key inside that run, e.g. `terminal_decision`.",
              "title": "Stage Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record a signed human decision",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This deployment",
      "url": "/api/v1"
    }
  ],
...
}